Updated unittests
